+++ core/src/lib/btimers.cc
@@ -188,9 +188,9 @@ btimer_t* StartBsockTimer(BareosSocket* bsock, uint32_t wait)
   RegisterWatchdog(wid->wd);
 
   Dmsg4(debuglevel,
-        "Start bsock timer %p tid=%s for %" PRIu32 " secs at %llu\n", wid,
-        edit_pthread(wid->tid, ed1, sizeof(ed1)), wait,
-        static_cast<long long unsigned>(time(NULL)));
+        "Start bsock timer %p tid=%s for %" PRIu32 " secs at %" PRIu64 "\n",
+        wid, edit_pthread(wid->tid, ed1, sizeof(ed1)), wait,
+        static_cast<uint64_t>(time(NULL)));
 
   return wid;
 }
@@ -205,9 +205,9 @@ void StopBsockTimer(btimer_t* wid)
     return;
   }
 
-  Dmsg3(debuglevel, "Stop bsock timer %p tid=%s at %llu.\n", wid,
+  Dmsg3(debuglevel, "Stop bsock timer %p tid=%s at %" PRIu64 ".\n", wid,
         edit_pthread(wid->tid, ed1, sizeof(ed1)),
-        static_cast<long long unsigned>(time(NULL)));
+        static_cast<uint64_t>(time(NULL)));
   StopBtimer(wid);
 }
 
@@ -232,12 +232,12 @@ static void CallbackThreadTimer(watchdog_t* self)
   char ed1[50];
   btimer_t* wid = (btimer_t*)self->data;
 
-  Dmsg4(debuglevel, "thread timer %p kill %s tid=%p at %llu.\n", self,
+  Dmsg4(debuglevel, "thread timer %p kill %s tid=%p at %" PRIu64 ".\n", self,
         wid->type == TYPE_BSOCK ? "bsock" : "thread",
         edit_pthread(wid->tid, ed1, sizeof(ed1)),
-        static_cast<long long unsigned>(time(NULL)));
+        static_cast<uint64_t>(time(NULL)));
   if (wid->jcr) {
-    Dmsg2(debuglevel, "killed JobId=%u Job=%s\n", wid->jcr->JobId,
+    Dmsg2(debuglevel, "killed JobId=%" PRIu32 " Job=%s\n", wid->jcr->JobId,
           wid->jcr->Job);
   }
