+++ core/src/stored/acquire.cc
@@ -435,7 +435,7 @@ DeviceControlRecord* AcquireDeviceForAppend(DeviceControlRecord* dcr)
     dev->rLock(true);
     BlockDevice(dev, BST_DOING_ACQUIRE);
     dev->Unlock();
-    Dmsg1(190, "jid=%u Do mount_next_write_vol\n", (uint32_t)jcr->JobId);
+    Dmsg1(190, "jid=%" PRIu32 " Do mount_next_write_vol\n", jcr->JobId);
     if (!dcr->MountNextWriteVolume()) {
       if (!jcr->IsJobCanceled()) {
         /* Reduce "noise" -- don't print if job canceled */
@@ -457,8 +457,8 @@ DeviceControlRecord* AcquireDeviceForAppend(DeviceControlRecord* dcr)
   if (jcr->sd_impl->NumWriteVolumes == 0) { jcr->sd_impl->NumWriteVolumes = 1; }
   dev->VolCatInfo.VolCatJobs++; /* increment number of jobs on vol */
   Dmsg4(100, "=== nwriters=%d nres=%d vcatjob=%" PRIu32 " dev=%s\n",
-        dev->num_writers,
-        dev->NumReserved(), dev->VolCatInfo.VolCatJobs, dev->print_name());
+        dev->num_writers, dev->NumReserved(), dev->VolCatInfo.VolCatJobs,
+        dev->print_name());
   dcr->DirUpdateVolumeInfo(
       is_labeloperation::False); /* send Volume info to Director */
   retval = true;
@@ -636,9 +636,8 @@ bool ReleaseDevice(DeviceControlRecord* dcr)
   }
 
   pthread_cond_broadcast(&dev->wait_next_vol);
-  Dmsg2(100, "JobId=%u broadcast wait_device_release at %s\n",
-        (uint32_t)jcr->JobId,
-        bstrftimes(tbuf, sizeof(tbuf), (utime_t)time(NULL)));
+  Dmsg2(100, "JobId=%" PRIu32 " broadcast wait_device_release at %s\n",
+        jcr->JobId, bstrftimes(tbuf, sizeof(tbuf), (utime_t)time(NULL)));
   ReleaseDeviceCond();
 
   // If we are the thread that blocked the device, then unblock it
@@ -656,8 +655,8 @@ bool ReleaseDevice(DeviceControlRecord* dcr)
     FreeDeviceControlRecord(dcr);
   }
 
-  Dmsg2(100, "Device %s released by JobId=%u\n", dev->print_name(),
-        (uint32_t)jcr->JobId);
+  Dmsg2(100, "Device %s released by JobId=%" PRIu32 "\n", dev->print_name(),
+        jcr->JobId);
 
   return retval;
 }
